refactor: 에러 권한 페이지 수정
